Moved xmflib and constants to lib
