+++ xmf2obj.py
@@ -20,7 +20,7 @@ import glob
 from io import BytesIO
 from PIL import Image, ImageDraw
 from lib.x4lib import get_config, require_python_version, ModUtilMixin
-from xmflib import XMFException, XMFHeader, XMFChunk, XMFMaterial, VERTEX, NORMAL, UV
+from lib.xmflib import XMFException, XMFHeader, XMFChunk, XMFMaterial, VERTEX, NORMAL, UV
 
 require_python_version(3, 6)
 logger = logging.getLogger('x4.' + __name__)
